+++ CHANGES.rst
@@ -16,6 +16,9 @@ Changelog
 
 **Fixed**
 
+- #716 Samples from inside Batch are not filtered correctly
+- #707 AR Add: Set default contact on client change
+- #700 Fix filtering by review state in aggregated list of analyses
 - #715 AR Rejection Fails when e-mail Notification is enabled
 - #709 Fix removal not possible of last non-verified Analysis in Manage Analysis View
 - #706 Filtering by Department is not working